+++ CHANGELOG.md
@@ -79,3 +79,5 @@ which block all requests data thread)
 * Update `BotLogger` behaviour
 * Fix problem with canceling of next scheduled job for `SchedulerPlugin`
 * Fix behaviour of GetSchedulesCommand when queue is empty
+* Add for items of `SmartChooser` plugin config options `minAge` and `maxAge` for prefilter
+posts which too old or too young